Fix removed fractions.gcd in Python >= 3.9
`fractions.gcd` was removed. The recommended alternative is `math.gcd`
but it does not work with integers (and some of the test cases use
floats).

Re-implement gcd by using Python 3.8 implementation, but add a step to
turn floats into fractions.

This isn't pretty but this gives less surprising results for the kind of
numbers that are expected to appear in JSON schemas :

```
>>> from fractions import Fraction, gcd
>>> gcd(0.1, 0.01)
1.734723475976807e-18
>>> gcd(Fraction(str(0.1)), Fraction(str(0.01)))
Fraction(1, 100)
>>> gcd(0.6, 0.4)
1.1102230246251565e-16
>>> gcd(Fraction(str(0.6)), Fraction(str(0.4)))
Fraction(1, 5)
>>> from fractions import Fraction, gcd
>>> gcd(0.1, 0.01)
1.734723475976807e-18
>>> gcd(Fraction(str(0.1)), Fraction(str(0.01)))
Fraction(1, 100)
>>> gcd(0.6, 0.4)
1.1102230246251565e-16
>>> gcd(0.1, 0.25)
2.7755575615628914e-17
>>> gcd(Fraction(str(0.1)), Fraction(str(0.25)))
Fraction(1, 20)
```